# Conflicts:
#	Documents/NetBeansProjects/Group3_MotorPH_PayrollPaymentSystem/nbproject/build-impl.xml
#	Documents/NetBeansProjects/Group3_MotorPH_PayrollPaymentSystem/nbproject/genfiles.properties
#	Documents/NetBeansProjects/Group3_MotorPH_PayrollPaymentSystem/nbproject/project.properties
#	Documents/NetBeansProjects/Group3_MotorPH_PayrollPaymentSystem/nbproject/project.xml
#	Documents/NetBeansProjects/Group3_MotorPH_PayrollPaymentSystem/src/group3_motorph_payrollpaymentsystem/Group3_MotorPH_PayrollPaymentSystem.java

+++ AOOP_MPH_PayrollSystem_Maven/src/main/java/model/Attendance_sp.java
@@ -11,7 +11,7 @@ public class Attendance_sp {
     private Time timeIn;
     private Time timeOut;
     private Double regularHoursCalc;
-    private Integer overtimeRate; 
+    private Double overtimeRate; 
     private Double overtimeHoursCalc;
     private Date overtimeUpdatedDate;
     private String lastName;
@@ -30,7 +30,7 @@ public class Attendance_sp {
             Time timeIn,
             Time timeOut,
             Double regularHoursCalc,
-            Integer overtimeRate,
+            Double overtimeRate,
             Double overtimeHoursCalc,
             Date overtimeUpdatedDate,
             String lastName,
@@ -104,11 +104,11 @@ public class Attendance_sp {
         this.regularHoursCalc = regularHoursCalc;
     }
 
-    public Integer getOvertimeRate() {
+    public Double getOvertimeRate() {
         return overtimeRate;
     }
 
-    public void setOvertimeRate(Integer overtimeRate) {
+    public void setOvertimeRate(Double overtimeRate) {
         this.overtimeRate = overtimeRate;
     }
